[All] Rename tools folder to utils
