+++ unittest/com.cpp
@@ -21,7 +21,7 @@
 #include "pinocchio/algorithm/crba.hpp"
 #include "pinocchio/algorithm/rnea.hpp"
 #include "pinocchio/algorithm/center-of-mass.hpp"
-#include "pinocchio/tools/timer.hpp"
+#include "pinocchio/utils/timer.hpp"
 #include "pinocchio/parsers/sample-models.hpp"
 
 #include <iostream>
